+++ libs/dot-components/src/lib/components/input-form-fields/InputFormFields.styles.ts
@@ -35,7 +35,7 @@ export const StyledTextField = styled(TextField)`
       .MuiSelect-select:focus {
         background-color: transparent;
       }
-      .Mui-error.${adornmentIconClassName} {
+      .Mui-error .${adornmentIconClassName} {
           color: ${theme.palette.error[500]};
       }
       .MuiSelect-icon {